lab_02

+++ geremi.py
@@ -11,7 +11,7 @@ else:
  
 valor = salario*porcentagem 
 valor_a = valor - salario
-porcen = int((porcentagem - 1) *100)
+porcen = (porcentagem - 1) *100
  
 print(f"O seu antigo salário era R$ {salario:.2f} e com o reajuste de {porcen:.0f}% ficaria R$ {valor:.2f}")
 print(f"valor acrecentado foi de: R$ {valor_a:.2f}")
